Listening/requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
filelock==3.16.1
fsspec==2024.12.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.5
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
pycparser==2.22
regex==2024.11.6
requests==2.32.3
sounddevice==0.5.1
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.67.1
triton==3.0.0
typing-extensions==4.12.2
urllib3==2.2.3
wavio==0.0.9
zipp==3.20.2
